+++ spec/cmdx/configuration_spec.rb
@@ -23,6 +23,11 @@ RSpec.describe CMDx::Configuration do
       expect(config.middlewares).to be_a(CMDx::MiddlewareRegistry)
       expect(config.middlewares).to be_empty
     end
+
+    it "initializes with default hooks registry" do
+      expect(config.hooks).to be_a(CMDx::HookRegistry)
+      expect(config.hooks).to be_empty
+    end
   end
 
   describe "#to_h" do
@@ -30,7 +35,7 @@ RSpec.describe CMDx::Configuration do
       result = config.to_h
 
       expect(result).to be_a(Hash)
-      expect(result.keys).to match_array(%i[logger middlewares task_halt batch_halt])
+      expect(result.keys).to match_array(%i[logger middlewares hooks task_halt batch_halt])
     end
 
     it "returns current attribute values" do
@@ -68,6 +73,15 @@ RSpec.describe CMDx::Configuration do
       config.middlewares = new_middlewares
       expect(config.middlewares).to eq(new_middlewares)
       expect(config.middlewares.size).to eq(1)
+    end
+
+    it "allows reading and writing hooks" do
+      new_hooks = CMDx::HookRegistry.new
+      new_hooks.register(:before_execution, :test_hook)
+
+      config.hooks = new_hooks
+      expect(config.hooks).to eq(new_hooks)
+      expect(config.hooks[:before_execution]).to eq([[[:test_hook], {}]])
     end
   end
 
@@ -112,6 +126,46 @@ RSpec.describe CMDx::Configuration do
       expect(config.middlewares.size).to eq(2)
       expect(config.middlewares[0]).to eq([middleware1, [], nil])
       expect(config.middlewares[1]).to eq([middleware2, ["arg"], nil])
+    end
+  end
+
+  describe "#hooks" do
+    it "allows adding hooks to the global registry" do
+      config.hooks.register(:before_execution, :test_hook)
+
+      expect(config.hooks.keys).to include(:before_execution)
+      expect(config.hooks[:before_execution]).to eq([[[:test_hook], {}]])
+    end
+
+    it "allows adding hooks with conditions" do
+      config.hooks.register(:on_failure, :alert_admin, if: :production?)
+
+      expect(config.hooks[:on_failure]).to eq([[[:alert_admin], { if: :production? }]])
+    end
+
+    it "allows adding multiple hooks for the same event" do
+      config.hooks.register(:on_success, :log_success)
+      config.hooks.register(:on_success, :track_metrics)
+
+      expect(config.hooks[:on_success].size).to eq(2)
+      expect(config.hooks[:on_success][0]).to eq([[:log_success], {}])
+      expect(config.hooks[:on_success][1]).to eq([[:track_metrics], {}])
+    end
+
+    it "allows adding hook instances" do
+      hook_instance = CMDx::Hook.new
+
+      config.hooks.register(:before_validation, hook_instance)
+
+      expect(config.hooks[:before_validation]).to eq([[[hook_instance], {}]])
+    end
+
+    it "allows adding proc hooks" do
+      proc_hook = proc { |_task, _hook_type| puts "Hook executed" }
+
+      config.hooks.register(:after_execution, proc_hook)
+
+      expect(config.hooks[:after_execution]).to eq([[[proc_hook], {}]])
     end
   end
 end